Separate the langid test into two separate test scripts
